pyproject.toml
[project]
name = "maro"
version = "0.1.0"
description = "Multi-Agent Requirements Orchestrator - Event-driven system for automated requirements gathering from Slack"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }

dependencies = [
    # Web framework
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",

    # AutoGen for multi-agent orchestration
    "pyautogen>=0.2.0",

    # Graph processing
    "networkx>=3.2",

    # Database
    "sqlalchemy>=2.0.0",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",

    # Redis
    "redis>=5.0.0",
    "rq>=1.16.0",

    # Slack integration
    "slack-bolt>=1.18.0",

    # Jira integration
    "atlassian-python-api>=3.41.0",

    # LLM clients
    "openai>=1.12.0",
    "anthropic>=0.18.0",
    "google-generativeai>=0.4.0",

    # Utilities
    "pydantic>=2.6.0",
    "pydantic-settings>=2.1.0",
    "httpx>=0.26.0",
    "tenacity>=8.2.0",
    "structlog>=24.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.2.0",
    "mypy>=1.8.0",
    "pre-commit>=3.6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --cov=src --cov-report=term-missing"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true
